add job4 &5

+++ jour1/job4.java
@@ -4,7 +4,7 @@ public class job4
 public static void main( String[] args)
 {Scanner scanner = new Scanner(System.in);
     System.out.println("nb a calculer :");
-    double nb = scanner.nextInt();
+    double nb = scanner.nextDouble();
     System.out.println("resultat du nb au carré " + nb*nb);
 
 }
